Changed all returns to int's as required

+++ le.h
@@ -4,7 +4,7 @@ public:
     Le(Expression* left, Expression* right): SubExpression(left, right)
     {
     }
-    double evaluate()
+    int evaluate()
     {
         if(left->evaluate() < right->evaluate()) {
