+++ minus.h
@@ -4,7 +4,7 @@ public:
     Minus(Expression* left, Expression* right): SubExpression(left, right)
     {
     }
-    double evaluate()
+    int evaluate()
     {
         return left->evaluate() - right->evaluate();
     }
